Added lists

+++ src/alloc/alloc.h
@@ -13,8 +13,8 @@ struct fuse_allocator_header
 {
     void *ptr;                          ///< A pointer to the memory block
     size_t size;                        ///< The size of the memory block, in bytes
-    uint32_t magic;                     ///< A magic number
-    uint32_t ref;                       ///< The reference count of the memory block
+    uint16_t magic;                     ///< A magic number
+    uint16_t ref;                       ///< The reference count of the memory block
     bool used;                          ///< Whether the memory block is used
     struct fuse_allocator_header *prev; ///< The previous memory block header, or NULL if this is the first memory block header
     struct fuse_allocator_header *next; ///< The next memory block header, or NULL if this is the last memory block header
